requirements.txt
Django
Pillow
django-ckeditor
djangorestframework
djangorestframework-simplejwt
drf-yasg
python-dotenv